Change return type of getNativeScreenDensityNATIVE to double
Devices may report values like `1.25`.

+++ backends/backend-web/src/main/java/com/github/xpenatan/gdx/teavm/backends/web/WebGraphics.java
@@ -370,7 +370,7 @@ public abstract class WebGraphics implements Graphics {
     }
 
     @JSBody(script = "return devicePixelRatio || 1;")
-    private static native int getNativeScreenDensityNATIVE();
+    private static native double getNativeScreenDensityNATIVE();
 
     @JSBody(script = "return screen.width;")
     private static native int getScreenWidthNATIVE();
